initial work on the TaxonomyRouting feature
this feature defines a part that, when attached to a type, provides a
TaxonomyField allowing for a single selection (at any level of the taxonomy).

The selected term from the TaxonomyField is then used to generate an
Autoroute on the item, using the path of the term and its ancestors,
combined with the item's slug.

This essentially turns the taxonomy into a sitemap.